+++ lib/resource/uiFastor/responsive_views/helper/ResponsiveFastorConstant.dart
@@ -1,0 +1,5 @@
+class ResponsiveFastorConstant {
+
+  static double widthChangeInCaseLandscape = 600;
+  static double heightChangeInCaseLandscape = 900;
+}