+++ lib/resource/uiFastor/responsive_views/landscape/shape_content_center_top/PortraitBodyToCenterPageInLandscapeResponsiveFastor.dart
@@ -1,0 +1,25 @@
+import 'package:fastor_app_ui_widget/fastor_app_ui_widget.dart';
+import 'package:fastor_app_ui_widget/resource/uiFastor/responsive_views/helper/ResponsiveFastorConstant.dart';
+import 'package:flutter/material.dart';
+
+class PortraitBodyToCenterPageInLandscapeResponsiveFastor extends StatelessWidget {
+
+  Widget body ;
+
+  PortraitBodyToCenterPageInLandscapeResponsiveFastor( { required this.body });
+
+  @override
+  Widget build(BuildContext context) {
+    var boxTheChild =  Container(
+     width: DeviceTools.isLandscape(context)?ResponsiveFastorConstant.widthChangeInCaseLandscape : null ,
+      child: body,
+      alignment: DeviceTools.isLandscape(context)?Alignment.center : null,
+    );
+
+    return Container(
+      width: DeviceTools.getWidth(context),
+      alignment: Alignment.center,
+      child: boxTheChild,
+    );
+  }
+}