+++ lib/resource/uiFastor/responsive_views/landscape/shape_fourm_submit/ForumPortraitToLandscapeResponsiveFastor.dart
@@ -1,0 +1,90 @@
+
+
+
+import 'package:fastor_app_ui_widget/fastor_app_ui_widget.dart';
+import 'package:fastor_app_ui_widget/fastor_app_ui_widget.dart';
+import 'package:fastor_app_ui_widget/resource/uiFastor/responsive_views/helper/PortraitToLandscapeUtilsResponsiveFastor.dart';
+ 
+import 'package:flutter/cupertino.dart';
+
+
+ /// Responsive to landscape
+class ForumPortraitToLandscapeResponsiveFastor     extends StatelessWidget {
+
+  Color colorBackground;
+  Widget body ;
+
+  ForumPortraitToLandscapeResponsiveFastor( { required this.body , required this.colorBackground});
+
+  @override
+  Widget build(BuildContext context) {
+    return wrapHeight(context, body, colorBackground);
+  }
+
+  //------------------------------------------------------------------ wrap height
+
+  static Widget wrapHeight(BuildContext context, Widget content, Color colorBackground ) {
+    //check default mobile
+    if( DeviceTools.isPortrait( context) ) {
+      return content;
+    }
+
+    //stack
+    var stack =  StackTemplate.AlignParentPage( context: context,
+        alignCenterTop: _cardViewWrap( context, content, colorBackground )
+    );
+
+    //margin to parent
+    return Container(child: stack,
+        margin: EdgeInsets.all(DSDimen.space_parent)
+    );
+  }
+
+  static Widget _cardViewWrap(BuildContext context, Widget content, Color colorBackground ) {
+
+    //card
+    return Container( child:  content,
+      width: PortraitToLandscapeUtilsResponsiveFastor.getDeviceWidthPortraitNormalWhileLandscapeCenter( context),
+      decoration: BoarderHelper.cardViewShadow(
+        colorBackground:  colorBackground,
+        colorLine: colorBackground,
+
+      )  ,
+    );
+  }
+
+
+}
+
+/**
+
+    //------------------------------------------------------------------ fixed height
+
+    static Widget fixedHeight(BuildContext context, Widget content, double height ) {
+    //check default mobile
+    if( DeviceTools.isPortrait( context ) ) {
+    return content;
+    }
+
+    var stack =  StackTemplate.AlignParentPage( context: context,
+    alignCenterTop: _cardviewFixedHeight( context, content,  height)
+    );
+
+    //margin to parent
+    return Container(child: stack,
+    margin: EdgeInsets.all(DSDimen.space_parent)
+    );
+    }
+
+
+    static Widget _cardviewFixedHeight( BuildContext context, Widget child, double height){
+
+    return CardViewTemplate.t(
+    child:   child  ,
+    width: PortraitToLandscapeUtilsResponsiveFastor.getDeviceWidthPortraitNormalWhileLandscapeCenter( context),
+    height: height,
+    radius_all: DSDimen.ds_size_corner_level_2,
+    elevationValue: DSDimen.elevation_level_2,
+    );
+    }
+ */